code_ok/requirements.txt
opencv-python==4.8.1.78
numpy==1.24.3
pandas==2.0.3
tqdm==4.65.0
mediapipe==0.10.8
face_recognition==1.3.0
pymongo==4.5.0
matplotlib==3.7.2
scikit-learn==1.3.0
pillow==10.0.0
requests==2.31.0
flask==2.3.3
werkzeug==2.3.7 